Fix jenkins compile failure
Jenkins just started complaining about a mismatch between
an sprintf format and its type.

+++ cfam_access.cpp
@@ -51,7 +51,7 @@ void writeReg(const std::unique_ptr<Target>& target,
         char msg[100];
         sprintf(msg, "writeCFAMReg: Failed seek for address 0x%X, "
                 "processor %d.  errno = %d",
-                address, target->getPos(), errno);
+                address, static_cast<int>(target->getPos()), errno);
         throw std::runtime_error(msg);
     }
 
@@ -62,7 +62,7 @@ void writeReg(const std::unique_ptr<Target>& target,
         char msg[100];
         sprintf(msg, "writeCFAMReg: Failed write to address 0x%X, "
                 "processor %d. errno = %d",
-                address, target->getPos(), errno);
+                address, static_cast<int>(target->getPos()), errno);
         throw std::runtime_error(msg);
     }
 }
@@ -80,7 +80,7 @@ cfam_data_t readReg(const std::unique_ptr<Target>& target,
         char msg[100];
         sprintf(msg, "readCFAMReg: Failed seek for address 0x%X, "
                 "processor %d.  errno = %d",
-                address, target->getPos(), errno);
+                address, static_cast<int>(target->getPos()), errno);
         throw std::runtime_error(msg);
     }
 
@@ -91,7 +91,7 @@ cfam_data_t readReg(const std::unique_ptr<Target>& target,
         char msg[100];
         sprintf(msg, "readCFAMReg: Failed read for address 0x%X, "
                 "processor %d. errno = %d",
-                address, target->getPos(), errno);
+                address, static_cast<int>(target->getPos()), errno);
         throw std::runtime_error(msg);
     }
